Change the way the output name is computed
We want something like roto001 as the default output name (set in the template)
and I expect it to be automatically incremented to roto002.
The default implementation just adds a digit at the end so the output becomes
roto0011 ...
So I changed the populate_initial_output_name function to figure out the
padding and increment properly if digits are present in the output.